4-11-2

+++ back/cxmMusic/src/main/java/com/cxm/cxmmusic/vo/mongo/SonglistStarNumber.java
@@ -32,7 +32,7 @@ public class SonglistStarNumber {
     public void addTimes() {
         this.times += 1;
     }
-    public void addTimes(Integer times) {
+    public void addTimes(Long times) {
         this.times += times;
     }
